Work around different handling of ampersands in Windows 11 system tray
Windows 11 handles ampersands in system tray icon tooltips differently from earlier versions of Windows. Microsoft has shown no interest in fixing the problem (or even accepting that the problem exists), therefore this adds a nasty `if <is windows 11>` to work around the problem.
