+++ foo_ui_columns/win32.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace cui::win32 {
+
+bool check_windows_10_build(DWORD build_number);
+bool is_windows_11_rtm_or_newer();
+
+} // namespace cui::win32
